tests cleanup
https://coveord.atlassian.net/browse/KIT-5096
