+++ packages/atomic/src/components/common/product-template/stencil-product-template-common.tsx
@@ -7,7 +7,7 @@ import {
 import {h} from '@stencil/core';
 import {tableElementTagName} from '../../search/atomic-table-result/table-element-utils';
 import {ItemTarget} from '../layout/display-options';
-import {isResultSectionNode} from '../layout/layout-sections';
+import {isResultSectionNode} from '../layout/item-layout-sections';
 
 export {makeMatchConditions} from './product-template-common';
